refactor(web): extract managed media subsystem
